CAMEL-23855: Add F8 AI prompt panel to TUI
